Push 3
Added Enemy Health, Directional Shooting, Combat

+++ Assets/Scripts/HealthSystem.cs
@@ -32,7 +32,7 @@ public class HealthSystem : MonoBehaviour
         return this.healthMax;
     }
 
-    public void damage(int damage)
+    public void damage(float damage)
     {
         if(gracePeriod <= 0)
         {
